Minor refactoring to use useFormikContext and changes to component structure
